Refactor the local engine to use an in process executor
This is work in progress. Review is submitted for feedbacks. Refactored
launch script to start the API and executor on the same process so the
transport for a oslo.messaging fake driver can be shared.  A transport
attribute is added to the abstract engine as a class attribute to reference
the transport object.

Change-Id: I4b8684ebded9eb993749f2503380e078087edd7d
Implements: blueprint mistral-inproc-executor